# nids/api/schemas.py
"""Pydantic request contracts used by the Flask REST API."""
from __future__ import annotations

from typing import Any

from pydantic import BaseModel, ConfigDict, Field


class PredictRequest(BaseModel):
    model_config = ConfigDict(extra="forbid")
    record: dict[str, Any]


class BatchPredictRequest(BaseModel):
    model_config = ConfigDict(extra="forbid")
    records: list[dict[str, Any]] = Field(min_length=1, max_length=500)
    include_local_explanations: bool = False